+++ Forge/src/main/java/net/potionstudios/biomeswevegone/forge/ForgePlatformHandler.java
@@ -73,7 +73,7 @@ public final class ForgePlatformHandler implements PlatformHandler {
 
 	@Override
 	public Supplier<MobBucketItem> createMobBucket(Supplier<EntityType<? extends Mob>> entity, Supplier<Fluid> fluid, Supplier<SoundEvent> sound) {
-		return () -> new MobBucketItem(entity, fluid, sound, new Item.Properties().stacksTo(1).component(DataComponents.BUCKET_ENTITY_DATA, CustomData.EMPTY).setId(BiomesWeveGone.key(Registries.ITEM, entity.get().getDescriptionId() + "_bucket")));
+		return () -> new MobBucketItem(entity, fluid, sound, new Item.Properties().stacksTo(1).component(DataComponents.BUCKET_ENTITY_DATA, CustomData.EMPTY).setId(BiomesWeveGone.key(Registries.ITEM, entity.get().toShortString() + "_bucket")));
 	}
 
 	@Override